Remove creation of XML syntax formatter inside classes with logic
