+++ Sources/Kysect.DotnetProjectSystem/SolutionModification/DotnetSolutionModifier.cs
@@ -1,25 +1,10 @@
 ﻿using Kysect.CommonLib.BaseTypes.Extensions;
 using Kysect.DotnetProjectSystem.Parsing;
-using Kysect.DotnetProjectSystem.Projects;
 using Kysect.DotnetProjectSystem.Xml;
 using Microsoft.Extensions.Logging;
-using Microsoft.Language.Xml;
 using System.IO.Abstractions;
 
 namespace Kysect.DotnetProjectSystem.SolutionModification;
-
-public interface IXmlProjectFileModifyStrategy<TSyntax>
-    where TSyntax : SyntaxNode
-{
-    IReadOnlyCollection<TSyntax> SelectNodeForModify(XmlDocumentSyntax document);
-    SyntaxNode ApplyChanges(TSyntax syntax);
-}
-
-public static class SolutionItemNameConstants
-{
-    public const string DirectoryBuildProps = "Directory.Build.props";
-    public const string DirectoryPackagesProps = "Directory.Packages.props";
-}
 
 public class DotnetSolutionModifier
 {
@@ -31,8 +16,7 @@ public class DotnetSolutionModifier
         string solutionPath,
         IFileSystem fileSystem,
         ILogger logger,
-        SolutionFileContentParser solutionFileParser,
-        XmlDocumentSyntaxFormatter xmlDocumentSyntaxFormatter)
+        SolutionFileContentParser solutionFileParser)
     {
         solutionPath.ThrowIfNull();
         fileSystem.ThrowIfNull();
@@ -45,13 +29,11 @@ public class DotnetSolutionModifier
         var directoryBuildPropsModifier = new DotnetPropsModifier(
             fileSystem.Path.Combine(fileInfo.Directory.FullName, SolutionItemNameConstants.DirectoryBuildProps),
             fileSystem,
-            logger,
-            xmlDocumentSyntaxFormatter);
+            logger);
         var directoryPackagePropsModifier = new DotnetPropsModifier(
             fileSystem.Path.Combine(fileInfo.Directory.FullName, SolutionItemNameConstants.DirectoryPackagesProps),
             fileSystem,
-            logger,
-            xmlDocumentSyntaxFormatter);
+            logger);
 
         string solutionFileContent = fileSystem.File.ReadAllText(solutionPath);
         IReadOnlyCollection<DotnetProjectFileDescriptor> projectFileDescriptors = solutionFileParser.ParseSolutionFileContent(solutionFileContent);
@@ -60,7 +42,7 @@ public class DotnetSolutionModifier
         foreach (DotnetProjectFileDescriptor projectFileDescriptor in projectFileDescriptors)
         {
             string projectFullPath = fileSystem.Path.Combine(fileInfo.Directory.FullName, projectFileDescriptor.ProjectPath);
-            var projectModifier = new DotnetProjectModifier(projectFullPath, fileSystem, logger, xmlDocumentSyntaxFormatter);
+            var projectModifier = new DotnetProjectModifier(projectFullPath, fileSystem, logger);
             bool supportModification = projectModifier.SupportModification();
             if (!supportModification)
                 logger.LogWarning("Project {Path} use legacy csproj format and will be skipped.", projectModifier.Path);
@@ -78,91 +60,12 @@ public class DotnetSolutionModifier
         DirectoryPackagePropsModifier = directoryPackagePropsModifier;
     }
 
-    public void Save()
+    public void Save(XmlDocumentSyntaxFormatter syntaxFormatter)
     {
-        DirectoryBuildPropsModifier.Save();
-        DirectoryPackagePropsModifier.Save();
+        DirectoryBuildPropsModifier.Save(syntaxFormatter);
+        DirectoryPackagePropsModifier.Save(syntaxFormatter);
 
         foreach (DotnetProjectModifier projectModifier in Projects)
-            projectModifier.Save();
-    }
-}
-
-public class DotnetProjectModifier
-{
-    private readonly XmlDocumentSyntaxFormatter _xmlDocumentSyntaxFormatter;
-
-    public string Path { get; }
-    public DotnetProjectFile Accessor { get; }
-
-    private readonly IFileSystem _fileSystem;
-
-    public DotnetProjectModifier(string path, IFileSystem fileSystem, ILogger logger, XmlDocumentSyntaxFormatter xmlDocumentSyntaxFormatter)
-    {
-        _xmlDocumentSyntaxFormatter = xmlDocumentSyntaxFormatter;
-        Path = path.ThrowIfNull();
-        _fileSystem = fileSystem.ThrowIfNull();
-
-        if (!fileSystem.File.Exists(path))
-            throw new ArgumentException($"Project file with path {path} was not found");
-
-        Accessor = DotnetProjectFile.Create(Path, _fileSystem, logger);
-    }
-
-    public bool SupportModification()
-    {
-        return Accessor.IsSdkFormat();
-    }
-
-    public void Save()
-    {
-        _fileSystem.File.WriteAllText(Path, Accessor.ToXmlString(_xmlDocumentSyntaxFormatter));
-    }
-}
-
-public class DotnetPropsModifier
-{
-    private readonly string _path;
-    private readonly IFileSystem _fileSystem;
-    private readonly Lazy<DotnetProjectFile> _fileAccessor;
-    private XmlDocumentSyntaxFormatter _xmlDocumentSyntaxFormatter;
-    public DotnetProjectFile Accessor => _fileAccessor.Value;
-
-    public DotnetPropsModifier(string path, IFileSystem fileSystem, ILogger logger, XmlDocumentSyntaxFormatter xmlDocumentSyntaxFormatter)
-    {
-        _path = path;
-        _fileSystem = fileSystem;
-        _xmlDocumentSyntaxFormatter = xmlDocumentSyntaxFormatter;
-        _fileAccessor = new Lazy<DotnetProjectFile>(() => DotnetProjectFile.Create(_path, _fileSystem, logger));
-    }
-
-    public void Save()
-    {
-        if (!_fileAccessor.IsValueCreated)
-            return;
-
-        _xmlDocumentSyntaxFormatter = new XmlDocumentSyntaxFormatter();
-        _fileSystem.File.WriteAllText(_path, _fileAccessor.Value.ToXmlString(_xmlDocumentSyntaxFormatter));
-    }
-}
-
-public class SetTargetFrameworkModifyStrategy(string value) : IXmlProjectFileModifyStrategy<XmlElementSyntax>
-{
-    public IReadOnlyCollection<XmlElementSyntax> SelectNodeForModify(XmlDocumentSyntax document)
-    {
-        document.ThrowIfNull();
-
-        return document
-            .GetNodesByName("TargetFramework")
-            .OfType<XmlElementSyntax>()
-            .ToList();
-    }
-
-    public SyntaxNode ApplyChanges(XmlElementSyntax syntax)
-    {
-        syntax.ThrowIfNull();
-
-        XmlTextSyntax content = SyntaxFactory.XmlText(SyntaxFactory.XmlTextLiteralToken(value, null, null));
-        return syntax.ReplaceNode(syntax.Content.Single(), content);
+            projectModifier.Save(syntaxFormatter);
     }
 }
