+++ Sources/Kysect.DotnetProjectSystem/SolutionModification/SolutionItemNameConstants.cs
@@ -1,0 +1,7 @@
+﻿namespace Kysect.DotnetProjectSystem.SolutionModification;
+
+public static class SolutionItemNameConstants
+{
+    public const string DirectoryBuildProps = "Directory.Build.props";
+    public const string DirectoryPackagesProps = "Directory.Packages.props";
+}
